Mouse Button Lock: fix C26451 found by a local build, document building
- dllmain.cpp: cast the move-cancel delta operands to 64-bit before the
  subtraction so the C++ Core Guidelines analysis (C26451, enforced as an
  error in PowerToys) is satisfied.
- Dev doc: add a "Building and debugging" section.

Verified with a local x64 Release build: version/logger/SettingsAPI and the
module all compile, the module links into PowerToys.MouseButtonLock.dll
exporting powertoy_create, and C++ code analysis is clean.

+++ src/modules/MouseUtils/MouseButtonLock/dllmain.cpp
@@ -534,8 +534,8 @@ void MouseButtonLock::CheckMoveCancel(ButtonLockState& st, ULONGLONG now, int ho
         return;
     }
 
-    const long long dx = pt.x - st.downPos.x;
-    const long long dy = pt.y - st.downPos.y;
+    const long long dx = static_cast<long long>(pt.x) - st.downPos.x;
+    const long long dy = static_cast<long long>(pt.y) - st.downPos.y;
     if (dx * dx + dy * dy > static_cast<long long>(pixels) * pixels)
     {
         st.moveCancelled = true;
